+++ node-one/src/main/java/be/uantwerpen/fti/nodeone/domain/NodeStructure.java
@@ -3,7 +3,6 @@ package be.uantwerpen.fti.nodeone.domain;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.stereotype.Component;
 
 @Getter
 @Setter
