+++ node-one/src/main/java/be/uantwerpen/fti/nodeone/service/NetworkService.java
@@ -3,18 +3,13 @@ package be.uantwerpen.fti.nodeone.service;
 import be.uantwerpen.fti.nodeone.config.NetworkConfig;
 import be.uantwerpen.fti.nodeone.domain.NextAndPreviousNode;
 import be.uantwerpen.fti.nodeone.domain.NodeStructure;
-import be.uantwerpen.fti.nodeone.domain.RegisterNodeRequest;
-import be.uantwerpen.fti.nodeone.controller.dto.NodeStructureDto;
-import be.uantwerpen.fti.nodeone.domain.NodeStructure;
 import be.uantwerpen.fti.nodeone.domain.RemoveNodeRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.ApplicationListener;
-import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.stereotype.Service;
 
+import java.io.DataOutputStream;
 import java.io.IOException;
 import java.net.*;
 
@@ -24,7 +19,6 @@ import java.net.*;
 public class NetworkService {
     private final NetworkConfig networkConfig;
     private final RestService restService;
-    private final TcpService tcpService;
     private final NodeStructure nodeStructure;
 
     @Async
@@ -58,17 +52,38 @@ public class NetworkService {
         //Request ip with id next node namingservice (REST)
         String NextIp = restService.requestNodeIpWithHashValue(next);
         //Send id previous to next (TCP)
-        tcpService.sendUpdatePrevious(NextIp, previous);
+        sendUpdatePrevious(NextIp, previous);
         //Request ip with id previous node namingservice (REST)
         String PreviousIp = restService.requestNodeIpWithHashValue(previous);
         //Send id next to previous (TCP)
-        tcpService.sendUpdateNext(PreviousIp, next);
+        sendUpdateNext(PreviousIp, next);
         restService.removeNode(new RemoveNodeRequest(hostname));
     }
 
     public void nodeFailure(String hostname) {
         NextAndPreviousNode nextAndPrevious = restService.getNextAndPrevious(hostname);
-        tcpService.sendUpdatePrevious(nextAndPrevious.getIpNext(), nextAndPrevious.getIdPrevious());
-        tcpService.sendUpdatePrevious(nextAndPrevious.getIpPrevious(), nextAndPrevious.getIdNext());
+        sendUpdatePrevious(nextAndPrevious.getIpNext(), nextAndPrevious.getIdPrevious());
+        sendUpdatePrevious(nextAndPrevious.getIpPrevious(), nextAndPrevious.getIdNext());
     }
+
+    public void sendUpdateNext(String ipAddress, int newNextNode) {
+        try (Socket socket = new Socket(ipAddress, 5000)) {
+            DataOutputStream outputStream = new DataOutputStream(socket.getOutputStream());
+            outputStream.writeInt(newNextNode);
+            outputStream.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public void sendUpdatePrevious(String ipAddress, int newPreviousNode) {
+        try (Socket socket = new Socket(ipAddress, 5000)) {
+            DataOutputStream outputStream = new DataOutputStream(socket.getOutputStream());
+            outputStream.writeInt(newPreviousNode);
+            outputStream.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
 }
